Factoring out common MOS parameters
