+++ SpiceSharp/Components/Semiconductors/MOS/Level3/ModelBaseParameters.cs
@@ -6,120 +6,93 @@ namespace SpiceSharp.Components.MosfetBehaviors.Level3
     /// <summary>
     /// Base parameters for a <see cref="Mosfet3Model"/>
     /// </summary>
-    public class ModelBaseParameters : ParameterSet
+    public class ModelBaseParameters : Common.ModelBaseParameters
     {
         /// <summary>
         /// Parameters
         /// </summary>
-        [ParameterName("vto"), ParameterName("vt0"), ParameterInfo("Threshold voltage")]
-        public GivenParameter<double> Vt0 { get; } = new GivenParameter<double>();
-        [ParameterName("kp"), ParameterInfo("Transconductance parameter")]
-        public GivenParameter<double> Transconductance { get; } = new GivenParameter<double>(2e-5);
-        [ParameterName("gamma"), ParameterInfo("Bulk threshold parameter")]
-        public GivenParameter<double> Gamma { get; } = new GivenParameter<double>();
-        [ParameterName("phi"), ParameterInfo("Surface potential")]
-        public GivenParameter<double> Phi { get; } = new GivenParameter<double>(.6);
-        [ParameterName("rd"), ParameterInfo("Drain ohmic resistance")]
-        public GivenParameter<double> DrainResistance { get; } = new GivenParameter<double>();
-        [ParameterName("rs"), ParameterInfo("Source ohmic resistance")]
-        public GivenParameter<double> SourceResistance { get; } = new GivenParameter<double>();
-        [ParameterName("cbd"), ParameterInfo("B-D junction capacitance")]
-        public GivenParameter<double> CapBd { get; } = new GivenParameter<double>();
-        [ParameterName("cbs"), ParameterInfo("B-S junction capacitance")]
-        public GivenParameter<double> CapBs { get; } = new GivenParameter<double>();
-        [ParameterName("is"), ParameterInfo("Bulk junction sat. current")]
-        public GivenParameter<double> JunctionSatCur { get; } = new GivenParameter<double>(1e-14);
-        [ParameterName("pb"), ParameterInfo("Bulk junction potential")]
-        public GivenParameter<double> BulkJunctionPotential { get; } = new GivenParameter<double>(.8);
-        [ParameterName("cgso"), ParameterInfo("Gate-source overlap cap.")]
-        public GivenParameter<double> GateSourceOverlapCapFactor { get; } = new GivenParameter<double>();
-        [ParameterName("cgdo"), ParameterInfo("Gate-drain overlap cap.")]
-        public GivenParameter<double> GateDrainOverlapCapFactor { get; } = new GivenParameter<double>();
-        [ParameterName("cgbo"), ParameterInfo("Gate-bulk overlap cap.")]
-        public GivenParameter<double> GateBulkOverlapCapFactor { get; } = new GivenParameter<double>();
-        [ParameterName("rsh"), ParameterInfo("Sheet resistance")]
-        public GivenParameter<double> SheetResistance { get; } = new GivenParameter<double>();
-        [ParameterName("cj"), ParameterInfo("Bottom junction cap per area")]
-        public GivenParameter<double> BulkCapFactor { get; } = new GivenParameter<double>();
-        [ParameterName("mj"), ParameterInfo("Bottom grading coefficient")]
-        public GivenParameter<double> BulkJunctionBotGradingCoefficient { get; } = new GivenParameter<double>(.5);
-        [ParameterName("cjsw"), ParameterInfo("Side junction cap per area")]
-        public GivenParameter<double> SidewallCapFactor { get; } = new GivenParameter<double>();
-        [ParameterName("mjsw"), ParameterInfo("Side grading coefficient")]
-        public GivenParameter<double> BulkJunctionSideGradingCoefficient { get; } = new GivenParameter<double>(.33);
-        [ParameterName("js"), ParameterInfo("Bulk jct. sat. current density")]
-        public GivenParameter<double> JunctionSatCurDensity { get; } = new GivenParameter<double>();
-        [ParameterName("tox"), ParameterInfo("Oxide thickness")]
-        public GivenParameter<double> OxideThickness { get; } = new GivenParameter<double>(1e-7);
-        [ParameterName("ld"), ParameterInfo("Lateral diffusion")]
-        public GivenParameter<double> LateralDiffusion { get; } = new GivenParameter<double>();
-        [ParameterName("u0"), ParameterName("uo"), ParameterInfo("Surface mobility")]
-        public GivenParameter<double> SurfaceMobility { get; } = new GivenParameter<double>(600);
-        [ParameterName("fc"), ParameterInfo("Forward bias jct. fit parm.")]
-        public GivenParameter<double> ForwardCapDepletionCoefficient { get; } = new GivenParameter<double>(.5);
-        [ParameterName("nsub"), ParameterInfo("Substrate doping")]
-        public GivenParameter<double> SubstrateDoping { get; } = new GivenParameter<double>();
-        [ParameterName("tpg"), ParameterInfo("Gate type")]
-        public GivenParameter<double> GateType { get; } = new GivenParameter<double>();
-        [ParameterName("nss"), ParameterInfo("Surface state density")]
-        public GivenParameter<double> SurfaceStateDensity { get; } = new GivenParameter<double>();
         [ParameterName("eta"), ParameterInfo("Vds dependence of threshold voltage")]
         public GivenParameter<double> Eta { get; } = new GivenParameter<double>();
-        [ParameterName("nfs"), ParameterInfo("Fast surface state density")]
-        public GivenParameter<double> FastSurfaceStateDensity { get; } = new GivenParameter<double>();
-        [ParameterName("theta"), ParameterInfo("Vgs dependence on mobility")]
-        public GivenParameter<double> Theta { get; } = new GivenParameter<double>();
-        [ParameterName("vmax"), ParameterInfo("Maximum carrier drift velocity")]
-        public GivenParameter<double> MaxDriftVelocity { get; } = new GivenParameter<double>();
-        [ParameterName("kappa"), ParameterInfo("Kappa")]
-        public GivenParameter<double> Kappa { get; } = new GivenParameter<double>(.2);
-        [ParameterName("xj"), ParameterInfo("Junction depth")]
-        public GivenParameter<double> JunctionDepth { get; } = new GivenParameter<double>();
-        [ParameterName("tnom"), DerivedProperty(), ParameterInfo("Parameter measurement temperature")]
-        public double NominalTemperatureCelsius
-        {
-            get => NominalTemperature - Circuit.CelsiusKelvin;
-            set => NominalTemperature.Value = value + Circuit.CelsiusKelvin;
-        }
-        public GivenParameter<double> NominalTemperature { get; } = new GivenParameter<double>(Circuit.ReferenceTemperature);
 
         /// <summary>
-        /// Methods
+        /// Gets the kappa parameter.
         /// </summary>
-        [ParameterName("delta"), ParameterInfo("Width effect on threshold")]
+        /// <value>
+        /// The kappa parameter.
+        /// </value>
+        [ParameterName("kappa"), ParameterInfo("Kappa")]
+        public GivenParameter<double> Kappa { get; } = new GivenParameter<double>(0.2);
+
+        /// <summary>
+        /// Gets the gate-source voltage dependence on mobility.
+        /// </summary>
+        /// <value>
+        /// The gate-source voltage dependence.
+        /// </value>
+        [ParameterName("theta"), ParameterInfo("Vgs dependence on mobility")]
+        public GivenParameter<double> Theta { get; } = new GivenParameter<double>();
+
+        /// <summary>
+        /// Gets the fast surface state density.
+        /// </summary>
+        /// <value>
+        /// The fast surface state density.
+        /// </value>
+        [ParameterName("nfs"), ParameterInfo("Fast surface state density")]
+        public GivenParameter<double> FastSurfaceStateDensity { get; } = new GivenParameter<double>();
+
+        /// <summary>
+        /// Gets the maximum drift velocity.
+        /// </summary>
+        /// <value>
+        /// The maximum drift velocity.
+        /// </value>
+        [ParameterName("vmax"), ParameterInfo("Maximum carrier drift velocity")]
+        public GivenParameter<double> MaxDriftVelocity { get; } = new GivenParameter<double>();
+
+        /// <summary>
+        /// Gets the junction depth.
+        /// </summary>
+        /// <value>
+        /// The junction depth.
+        /// </value>
+        [ParameterName("xj"), ParameterInfo("Junction depth")]
+        public GivenParameter<double> JunctionDepth { get; } = new GivenParameter<double>();
+
+        /// <summary>
+        /// Gets or sets the width effect on the threshold voltage.
+        /// </summary>
+        /// <value>
+        /// The narrowing factor.
+        /// </value>
+        /// <remarks>
+        /// When setting the parameter, spice 3f5 would change the delta parameter, but when asking
+        /// for it, Spice 3f5 would return the narrow factor instead. This behavior is copied here
+        /// for compatibility.
+        /// </remarks>
+        [ParameterName("delta"), ParameterInfo("Width effect on the threshold voltage")]
         public double DeltaWidth
         {
             get => NarrowFactor;
             set => Delta = value;
         }
-        [ParameterName("nmos"), ParameterInfo("N type Mosfet model")]
-        public void SetNmos(bool value)
-        {
-            if (value)
-                MosfetType = 1.0;
-        }
-        [ParameterName("pmos"), ParameterInfo("P type Mosfet model")]
-        public void SetPmos(bool value)
-        {
-            if (value)
-                MosfetType = -1.0;
-        }
-        [ParameterName("type"), ParameterInfo("N-channel or P-channel MOS")]
-        public string TypeName
-        {
-            get
-            {
-                if (MosfetType > 0)
-                    return "nmos";
-                return "pmos";
-            }
-        }
 
+        /// <summary>
+        /// Gets or sets the width effect on the threshold voltage.
+        /// </summary>
+        /// <value>
+        /// The delta.
+        /// </value>
         [ParameterName("input_delta"), ParameterInfo("")]
         public double Delta { get; protected set; }
+
+        /// <summary>
+        /// Gets or sets the narrowing factor.
+        /// </summary>
+        /// <value>
+        /// The narrowing factor.
+        /// </value>
         public double NarrowFactor { get; set; }
-        public double MosfetType { get; protected set; } = 1.0;
-        public double OxideCapFactor { get; private set; }
 
         /// <summary>
         /// Constructor
@@ -145,12 +118,33 @@ namespace SpiceSharp.Components.MosfetBehaviors.Level3
         /// </summary>
         public override void CalculateDefaults()
         {
-            OxideCapFactor = 3.9 * 8.854214871e-12 / OxideThickness;
-            if (!Transconductance.Given)
-                Transconductance.RawValue = SurfaceMobility * OxideCapFactor * 1e-4;
+            // Set the default oxide thickness
+            if (!OxideThickness.Given)
+                OxideThickness.RawValue = 1e-7;
 
-            // now model parameter preprocessing
+            // Calculate base defaults
+            base.CalculateDefaults();
+
+            // Calculate the narrowing factor
             NarrowFactor = Delta * 0.5 * Math.PI * Transistor.EpsilonSilicon / OxideCapFactor;
+        }
+
+        /// <summary>
+        /// Creates a deep clone of the parameter set.
+        /// </summary>
+        /// <returns>
+        /// A deep clone of the parameter set.
+        /// </returns>
+        public override ParameterSet DeepClone()
+        {
+            // We have some private/protected properties that need to be set manually.
+            var result = (ModelBaseParameters) base.DeepClone();
+
+            // Set properties
+            result.Delta = Delta;
+            result.NarrowFactor = NarrowFactor; // Just to be sure, given the special nature of "delta"
+
+            return result;
         }
     }
 }
